\name{rochdale}
\Rdversion{1.0.1}
\alias{rochdale}
\docType{data}
\title{The rochdale data}
\description{An 8 way sparse contingency table representing the cross classification of N=665 individuals. All 8 classification criteria are binary. The study was conducted to elicit information about factors affecting the pattern of economic life in Rochdale, England. The table is sparse have 165 counts of zero, 217 counts with at most three observations, but also a few large counts with 30 or more observations. The variables are: 
a. wife economically active (no, yes); 
b. age of wife >38 (no, yes); 
c. husband unemployed (no, yes); 
d. child <=4 (no, yes); 
e. wife's education, highschool+ (no, yes); 
f. husband's education, highschool+ (no, yes); 
g. Asian origin (no, yes); 
h. other household member working (no, yes). 
}
\usage{data(rochdale)}
\source{Whittaker (1990)} 
\references{[1] Whittaker, J. (1990). Graphical models in applied multivariate statistics. John Wiley & Sons.}
